requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
pdfplumber==0.11.4
tiktoken==0.8.0
sentence-transformers==3.3.1
numpy==2.1.3
faiss-cpu==1.9.0
pip install rank-bm25==0.2.2
groq==0.13.0
matplotlib==3.10.0
pandas==2.2.3
google-generativeai==0.8.3